+++ src/articles/articles.service.ts
@@ -2,7 +2,7 @@ import { Injectable } from '@nestjs/common';
 import { Repository } from 'typeorm';
 import { Article } from './entities/articles.entity';
 import { InjectRepository } from '@nestjs/typeorm';
-import { SaveArticleDto } from './articles.dto';
+import { SaveArticleDto } from './dto/articles.dto';
 
 @Injectable()
 export class ArticlesService {
